requirements.txt
black
fastapi
uvicorn[standard]
langchain
openai
python-dotenv
python-multipart